Keep role slugs in sync with roles in WP_Roles.
If a role no longer exists, we don't need to show or store the corresponding slug.

Previously, we used the `editable_roles` filter, which effectively hid some roles. This was not ideal. It also created a song and dance to keep everything in sync. This commit also addresses these issues.