Revert options back-compat deprecation
Part of the `BA_Edit_Author_Slug::options_back_compat()` method was introduced is because we needed `ba_eas_sanitize_author_base()` to be available. This meant that setting properties to the values in the db needed to happen after `BA_Edit_Author_Slug::includes()`.

Since the method needs to be reverted, we might as well also just keep the back-compat code too.

A later commit will perform some cleanup and split the options code into smaller methods.