Build Tools: Use master branch of `eslint-plugin-wordpress`
Until an official version is released with these changes, we'll use the master branch of https://github.com/WordPress-Coding-Standards/eslint-plugin-wordpress.

The majority of the rules seem to be built from `eslint-plugin-wpcalypso`, which doesn't enforce Yoda conditions. So, we add an extra rule to enforce them. This is the only obvious rule I've seen missing thus far.

Fixes #5