Tests: Optimize test for `BA_Edit_Author_Slug::load_textdomain`
The assertion on the return of `BA_Edit_Author_Slug::load_textdomain` fails on WP 4.4 and 4.5. It's not necessary for our application.

The extra test for unloading the text domain is, again, not necessary for our application.

In both cases, WP has done this testing for us. No need to duplicate. As long as our text domain is loaded, we're good to go.