Refactor plugin bootstrap, Part 1
The main goal of this initial refactoring is to do the minimum amount of work required to get `BA_Edit_Author_Slug` into it's own file. The plan for now is to conitnue the refactoring, but it will require deeper changes which I wasn't ready to risk. After the test suite is refactored, I'll feel more comfortable making additional changes to the bootstrap.