Tests: Fix `test_ba_eas_show_user_nicename_scripts()`
When not on the page `profile.php` we should not be registering/loading our JS. The prior iteration of this test wasn't properly checking this.

First, we were actually testing that it was registered. Second, by testing `profile.php` first, the script was registered/loaded, so that even if we used the correct assertion, the test would fail.