Reduce potential churn and just hash the user id and login
The idea behind the previous version was to reduce potential collisions. Let's say we have a user with an id of `1` and a login of `1guy`. Later we get a user with an id of `11` and a login of `guy`. This would cause a collision with the hash for the first user. So, I added the email address to the data being hashed. However, this has the potential to increase changes to the hash that would be reflected in the UI. Since user id and login rarely ever change, the easy thing to do would be to add a separator between the two. This reduced the likelyhood of a collision, and removes the need for the email address. The possibility that the UI could reflect a hash change still remains, but I have a feeling that people rarely rotate their salts in their wp-config.php, so this shouldn't be a major factor.