Switch to using a standard md5 hash
We're not going for crytographic security here. On top of that, iThemes Security encourages the regular rotation of the salts, and makes it easy. This potential churn is not a risk I'm willing to take, nor is it something I want to spend time answering questions about why it mysteriously changed.